backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pymongo==4.5.0
redis==5.0.4
openai==1.54.4
anthropic==0.40.0
pydantic>=2.6.4
pydantic-settings>=2.0.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.2.0
aiofiles>=23.0.0
pillow>=10.0.0
websockets>=12.0
slowapi>=0.1.9
tenacity>=8.2.3
structlog>=24.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.25.0
faker>=26.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
boto3>=1.34.129
requests>=2.31.0
tiktoken>=0.7.0
markdown>=3.6
bleach>=6.1.0